streamline data flow for UI mounting & evaluation form
The RSVP reader now shows the question as part of the text,
and shows the yes/no options immediately afterwards on the same page.
This also avoids the need to encode the statistics to pass them to the `/form`
page, which could be removed completely.

As part of the UI modifications, the reader view now uses CSS grids. This allows
the "message" field to be shown directly under the word, without the UI jumping
around when the message is shown/hidden. This way the reader can also be mounted
into a tutorial page easily.