remember last manual speed, extract state management
When the user manually adjusts the speed, that speed will be remembered and used
as initial speed for the next text where we don't prescribe a speed via manual
adjustment.

The state management with the cookies was getting out of hand. Now there's a
`UserState` class that contains corresponding fields, accessors, and state
transitions. It can be converted to/from the cookies.