tokenize text within RSVP -> hyphenation, pause after sentence
By tokenizing the text within the RSVP class, more intelligent tokenization is
possible.

First, it seems sensible to pause after sentences similar to how Spritz works.
This is currently done in a brute-force way by repeating the token. Works well,
except that fast-forward/rewind will be broken.

Second, we have a couple of very long hyphenated words that are hard to read.
RSVP doesn't work well beyond maybe 12 characters, and definitely breaks down at
20 chars. Using a regex we can break up long words:

    [African-American] -> [African-] [-American]

Note that the hyphen is included in both tokens to emphasize the connection.

Future work would be to move to the next token via `setTimeout()` rather than
`setInterval()` so that tokens can be shown for a different number of
milliseconds. This would enable removal of the sentence-pause hack.